% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{bmrb_from_pdb}
\alias{bmrb_from_pdb}
\title{Get BMRB IDs associated with a PDB entry}
\usage{
bmrb_from_pdb(pdb_id)
}
\arguments{
\item{pdb_id}{A PDB ID string (e.g. \code{"2JM0"}).}
}
\value{
A data.frame with columns \code{bmrb_id}, \code{match_type},
  and \code{comment}.
}
\description{
Get BMRB IDs associated with a PDB entry
}
\examples{
\dontrun{
bmrb_from_pdb("2JM0")
}
}
