% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{bmrb_pdb_ids}
\alias{bmrb_pdb_ids}
\title{Get PDB IDs associated with a BMRB entry}
\usage{
bmrb_pdb_ids(bmrb_id)
}
\arguments{
\item{bmrb_id}{A single BMRB entry ID.}
}
\value{
A data.frame with columns \code{pdb_id}, \code{match_type},
  and \code{comment}.
}
\description{
Get PDB IDs associated with a BMRB entry
}
\examples{
\dontrun{
bmrb_pdb_ids(15000)
}
}
