% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{chemical_shift_corr_res}
\alias{chemical_shift_corr_res}
\title{Chemical shift correlation between given pair of atoms in a given amino acid (or) nucleic acid}
\usage{
chemical_shift_corr_res(atom1, atom2, res)
}
\arguments{
\item{atom1}{atom name in NMR-STAR nomenclature like CA,CB2}

\item{atom2}{atom name in NMR-STAR nomenclature like HA,HB2}

\item{res}{residue name in NMR-STAR nomenclature like ALA}
}
\value{
plot object
}
\description{
Plots the correlated chemical shift distribution of given pair of atoms in a single residue  from BMRB database.
By default it will generate interactive graphics using plotly library
}
\examples{
plt<-chemical_shift_corr_res('HE21','HE22')
#plots the chemical shift distribution between HE21 and HE22
}
\seealso{
\code{\link{fetch_atom_chemical_shifts}}
}

