% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{convert_cs_to_n15hsqc}
\alias{convert_cs_to_n15hsqc}
\title{Converts the output data frame of fetch_entry_chemical_shifts into H1-N15 HSQC data frame}
\usage{
convert_cs_to_n15hsqc(csdf)
}
\arguments{
\item{csdf}{==> chemical shift data frame from fetch_entry_chemical_shift}
}
\value{
1H-N15 chemical shift list on the same row combined using comp index ID and bmrb ID
}
\description{
Converts the output data frame of fetch_entry_chemical_shifts into H1-N15 HSQC data frame
}
\examples{
df<-fetch_entry_chemical_shifts(15060)
hsqc<-convert_cs_to_n15hsqc(df)
}

