% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{convert_cs_to_tocsy}
\alias{convert_cs_to_tocsy}
\title{Reformats chemical shift dataframe for easy plotting}
\usage{
convert_cs_to_tocsy(csdf)
}
\arguments{
\item{csdf}{==> chemical shift data frame from fetch_entry_chemical_shift}
}
\value{
Tow 1H chemical shift lists on the same row combined using comp index ID and bmrb ID
}
\description{
Reformats the output dataframe from fetch_entry_chemical_shifts() into a simple dataframe contains only proton shifts, which could be used to simulate TOCSY spectra
}
\examples{
df<-fetch_entry_chemical_shifts(15060)
hsqc<-convert_cs_to_tocsy(df)
}

