% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cs_stats}
\alias{cs_stats}
\title{Compute summary statistics for a set of chemical shifts}
\usage{
cs_stats(x)
}
\arguments{
\item{x}{Numeric vector of chemical shift values.}
}
\value{
A named numeric vector: n, mean, sd, median, min, max.
}
\description{
Compute summary statistics for a set of chemical shifts
}
\examples{
cs_stats(c(55.1, 55.3, 56.0, 54.8))
}
