% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical_shift.R
\name{cs_table}
\alias{cs_table}
\title{Get a flat data.frame of chemical shifts for one or more BMRB entries}
\usage{
cs_table(bmrb_ids, auth_tag = FALSE)
}
\arguments{
\item{bmrb_ids}{A single BMRB entry ID or a vector of IDs.}

\item{auth_tag}{Logical. Use `Auth_seq_ID` for sequence numbering. Default FALSE.}
}
\value{
A data.frame suitable for use with ggplot2 or dplyr.
}
\description{
Get a flat data.frame of chemical shifts for one or more BMRB entries
}
\examples{
\dontrun{ df <- cs_table(c(15060, 17076)) }
}
