% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{fetch_atom_chemical_shift}
\alias{fetch_atom_chemical_shift}
\title{Downloads the chemical shift table for a given atom from macromolecule/metabolomics database}
\usage{
fetch_atom_chemical_shift(atom, db)
}
\arguments{
\item{atom}{==> atom name like CA,CB2}

\item{db}{==> macromolecules, metabolomics}
}
\value{
list of all atom chemical shifts for all BMRB entries as a R data frame
}
\description{
Downloads the chemical shift table for a given atom from macromolecule/metabolomics database
}
\examples{
df<-fetch_atom_chemical_shift('CB2','macromolecules')
df<-fetch_atom_chemical_shift('C1','metabolomics')
}

