% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{fetch_atom_chemical_shifts}
\alias{fetch_atom_chemical_shifts}
\title{NMR Chemical shifts list for a given atom from BMRB}
\usage{
fetch_atom_chemical_shifts(atom, db = "macromolecules")
}
\arguments{
\item{atom}{atom name like CA,CB2}

\item{db}{macromolecules, metabolomics}
}
\value{
R data frame that contains full chemical shift list for a given atom
}
\description{
Downloads the full list of chemical shifts from BMRB(www.bmrb.wisc.edu) macromolecular/metabolomics database for a given atom
}
\examples{
df<-fetch_atom_chemical_shifts('CB2','macromolecules')
# Downloads CB2 chemical shifts from macromolecules database at BMRB
df<-fetch_atom_chemical_shifts('C1','metabolomics')
# Downloads C1 chemical shifts from metabolomics database at BMRB
}
\seealso{
\code{\link{fetch_entry_chemical_shifts}}
}

