% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{fetch_entry_cs}
\alias{fetch_entry_cs}
\title{Imports checmial shift table for a given entry id from BMRB data base}
\usage{
fetch_entry_cs(BMRBid)
}
\arguments{
\item{BMRBid}{sinlge BMRB ID
For metabolomics entries entry id should have 'bmse' prefix example: c('bmse000034','bmse000035','bmse000036')}
}
\value{
R data frame that contains  Atom_chem_shift data for a given list of entries
}
\description{
Downloads NMR chemical shift data from BMRB database for a given Entry ID
}
\examples{
df<-fetch_entry_cs(15060)
# Downloads NMR chemical shifts of a single entry from BMRB
df<-fetch_entry_cs('bmse000034')
# Downloads data from BMRB metabolomics database
}
\seealso{
\code{\link{fetch_entry_chemical_shifts}},\code{\link{fetch_atom_chemical_shifts}} and \code{\link{fetch_res_chemical_shifts}}
}
