% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{fetchallBMRB}
\alias{fetchallBMRB}
\title{Reads the full chemical shift csv frile from BMRB ftp site. Not recommended unless you have high speed internet connection}
\usage{

  fetchallBMRB(csvpath = "http://www.bmrb.wisc.edu/ftp/pub/bmrb/relational_tables/nmr-star3.1/Atom_chem_shift.csv")
}
\arguments{
\item{Specify}{the csv file downloaded from http://www.bmrb.wisc.edu/ftp/pub/bmrb/relational_tables/nmr-star3.1/Atom_chem_shift.csv  with full path;May leave it empty to fetch the file directly from internet}
}
\value{
list of all atom chemical shifts for all BMRB entries as a R data frame
}
\description{
Reads the full chemical shift csv frile from BMRB ftp site. Not recommended unless you have high speed internet connection
}
\examples{
df<-fetchallBMRB()
df<-fetchallBMRB('~/Downloads/Atom_chem_shift.csv')
}

