% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{fetchBMRB}
\alias{fetchBMRB}
\title{Downloads chemical shift data from BMRB for a given BMRB entry/list of BMRB entries}
\usage{
fetchBMRB(BMRBidlist)
}
\arguments{
\item{BMRBidlist}{==> sinlge BMRB ID / list of BMRB IDs in csv format}
}
\value{
all available chemical shift data in R data frame
}
\description{
Downloads chemical shift data from BMRB for a given BMRB entry/list of BMRB entries
}
\examples{
df<-fetchBMRB(15060)
df<-fetchBMRB(c(15060,15070,8898,99))
}

