% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{fetchCSList}
\alias{fetchCSList}
\title{Reads the full chemical shift csv frile from BMRB ftp site. Not recommended unless you have high speed internet connection}
\usage{
fetchCSList(atom, db)
}
\arguments{
\item{atom}{==> atom name like CA,CB2}

\item{db}{==> macromolecules, metabolomics}
}
\value{
list of all atom chemical shifts for all BMRB entries as a R data frame
}
\description{
Reads the full chemical shift csv frile from BMRB ftp site. Not recommended unless you have high speed internet connection
}
\examples{
df<-fetchCSList('CB2','macromolecules')
df<-fetchCSList('C1','metabolomics')
}

