% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_cs_outliers}
\alias{filter_cs_outliers}
\title{Remove outliers from a numeric vector}
\usage{
filter_cs_outliers(x, sd_limit = 10)
}
\arguments{
\item{x}{Numeric vector.}

\item{sd_limit}{Number of standard deviations beyond which to remove.}
}
\value{
Numeric vector with outliers removed.
}
\description{
Remove outliers from a numeric vector
}
\examples{
filter_cs_outliers(c(52, 55, 56, 54, 300), sd_limit = 2)
}
