% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{filter_outlier}
\alias{filter_outlier}
\title{Remove chemical shift outliers}
\usage{
filter_outlier(cs = NA, cutoff = 8)
}
\arguments{
\item{cs}{data frame with amino acid information in Comp_ID and Atom_ID column}

\item{cutoff}{cutoff value(cutoff times standard deviation is used to trim the value on boths sides of mean)}
}
\value{
R data frame with chemical shift values
}
\description{
Removes chemical shifts values outside of cutoff times standard deviation on both sides of the mean
}
\examples{
#df<-filter_outlier(fetch_atom_chemical_shifts("CG2"))
#Downloads all CG2 chemical shifts and removes the outliers
}
\seealso{
\code{\link{filter_residue}} and \code{\link{fetch_atom_chemical_shifts}}
}
