% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{filter_residue}
\alias{filter_residue}
\title{Filter for standard 20 amino acids}
\usage{
filter_residue(df)
}
\arguments{
\item{df}{==> data frame with Comp_ID column}
}
\value{
filtered data frame
}
\description{
Filters 20 standard amino acids using Comp_ID
}
\examples{
df<-filter_residue(fetch_atom_chemical_shifts("CG2"))
}

