% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{HSQC_15N}
\alias{HSQC_15N}
\title{Simulate H1-N15 HSQC spectra}
\usage{
HSQC_15N(idlist, type = "scatter", interactive = TRUE)
}
\arguments{
\item{idlist}{==> list of bmrb ids c(17074,17076,17077)}

\item{type}{==> scatter/line default=scatter}

\item{interactive}{==> TRUE/FALSE default=TRUE}
}
\value{
plot object
}
\description{
Simulates H1-N15 HSQC spectra directly from BMRB(www.bmrb.wisc.edu) database. Default plot type will be 'scatter'.Peaks from different spectra(entries) can be connected based on residue numbers by specifying plot type as 'line'
}
\examples{
plot_hsqc<-HSQC_15N(c(17074,17076,17077))
plot_hsqc<-HSQC_15N(18857,'line')
plot_hsqc<-HSQC_15N(c(17074,17076,17077),interactive=FALSE)
}

