% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{makeRandomString}
\alias{makeRandomString}
\title{Generates random string of fixed length(for internal use in RBMRB)}
\usage{
makeRandomString()
}
\description{
Local files may not have Entry_ID, in that case random Entry_ID is assigned using this function. It is an internal function used only by RBMRB package
}

