% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{plot_n15hsqc}
\alias{plot_n15hsqc}
\title{Plots 15N-HSQC spectrum/spectra of given bmrb id/ list of bmrb ids. Plot type can be set as eight scatter (or) line.}
\usage{
plot_n15hsqc(idlist, type = "scatter")
}
\arguments{
\item{idlist}{==> list of bmrb ids c(17074,17076,17077)}

\item{type}{==> scatter/line default=scatter}
}
\value{
plot object
}
\description{
Plots 15N-HSQC spectrum/spectra of given bmrb id/ list of bmrb ids. Plot type can be set as eight scatter (or) line.
}
\examples{
plot_hsqc<-plot_n15hsqc(c(17074,17076,17077))
plot_hsqc<-plot_n15hsqc(18857,'line')
}

