% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{RBMRB_version}
\alias{RBMRB_version}
\title{Get the installed RBMRB version}
\usage{
RBMRB_version()
}
\value{
A character string with the package version.
}
\description{
Returns the version of the installed RBMRB package as a string.
}
\examples{
RBMRB_version()
}
