% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{simulate_c13hsqc}
\alias{simulate_c13hsqc}
\title{Simulate H1-C13 HSQC spectra}
\usage{
simulate_c13hsqc(idlist, type = "scatter", interactive = FALSE)
}
\arguments{
\item{idlist}{==> list of bmrb ids c(17074,17076,17077)}

\item{type}{==> scatter/line default=scatter}

\item{interactive}{==> TRUE/FALSE default=FALSE}
}
\value{
plot object
}
\description{
Simulates H1-C13 HSQC spectra directly from BMRB(www.bmrb.wisc.edu) database. Default plot type will be 'scatter'.Peaks from different spectra(entries) can be connected based on residue numbers by specifying plot type as 'line'
}
\examples{
plot_hsqc<-simulate_c13hsqc(c(17074,17076,17077))
plot_hsqc<-simulate_c13hsqc(18857,'line')
plot_hsqc<-simulate_c13hsqc(c(17074,17076,17077),interactive=TRUE)
}

