% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{simulate_n15hsqc}
\alias{simulate_n15hsqc}
\title{Simulate H1-N15 HSQC spectra}
\usage{
simulate_n15hsqc(idlist, type = "scatter", interactive = FALSE)
}
\arguments{
\item{idlist}{==> list of bmrb ids c(17074,17076,17077)}

\item{type}{==> scatter/line default=scatter}
}
\value{
plot object
}
\description{
Simulates H1-N15 HSQC spectra directly from BMRB(www.bmrb.wisc.edu) database. Default plot type will be 'scatter'.Peaks from different spectra(entries) can be connected based on residue numbers by specifying plot type as 'line'
}
\examples{
plot_hsqc<-simulate_n15hsqc(c(17074,17076,17077))
plot_hsqc<-simulate_n15hsqc(18857,'line')
}

