% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standard_amino_acids}
\alias{standard_amino_acids}
\title{List the 20 standard amino acid codes}
\usage{
standard_amino_acids()
}
\value{
Character vector of 3-letter amino acid codes.
}
\description{
List the 20 standard amino acid codes
}
\examples{
standard_amino_acids()
}
