% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standard_nucleic_acids}
\alias{standard_nucleic_acids}
\title{List standard nucleic acid residue codes}
\usage{
standard_nucleic_acids()
}
\value{
Character vector of RNA and DNA residue codes.
}
\description{
List standard nucleic acid residue codes
}
\examples{
standard_nucleic_acids()
}
