% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{chemical_shift_corr}
\alias{chemical_shift_corr}
\title{NMR 2D peak statistics}
\usage{
chemical_shift_corr(atom1, atom2, res = NA)
}
\arguments{
\item{atom1}{==> atom name like CA,CB2}

\item{atom2}{==> atom name like HA,HB2}

\item{res}{==> residue name like ALA,GLY}
}
\value{
plot object
}
\description{
Plots the correlated chemical shift distribution of any two atoms from BMRB database
}
\examples{
plt<-chemical_shift_corr('CG','HG2','MET')
}

