% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{create_2d_peaklist}
\alias{create_2d_peaklist}
\title{Build a generic 2D peak list for any two atoms}
\usage{
create_2d_peaklist(
  atom_x,
  atom_y,
  bmrb_ids = NULL,
  input_file_names = NULL,
  auth_tag = FALSE,
  draw_trace = FALSE,
  include_preceding = FALSE,
  include_next = FALSE
)
}
\arguments{
\item{atom_x}{Atom name for the X axis.}

\item{atom_y}{Atom name for the Y axis.}

\item{bmrb_ids}{BMRB entry ID(s) -- integer, character, or vector.}

\item{input_file_names}{Path(s) to local NMR-STAR `.str` files.}

\item{auth_tag}{Use `Auth_seq_ID` for sequence numbering.}

\item{draw_trace}{Draw lines connecting same-residue peaks across
datasets.}

\item{include_preceding}{Include preceding (i-1) residue Y-axis peaks.}

\item{include_next}{Include following  (i+1) residue Y-axis peaks.}
}
\value{
Same structure as [create_n15hsqc_peaklist()].
}
\description{
Equivalent to `pybmrb.Spectra.create_2d_peaklist()`.
}
