% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cs_peek}
\alias{cs_peek}
\title{Peek at a nested CS structure}
\usage{
cs_peek(cs_data, n_res = 5L)
}
\arguments{
\item{cs_data}{Output of [cs_from_bmrb()] or [cs_from_file()].}

\item{n_res}{Number of residues to show per chain.}
}
\description{
Peek at a nested CS structure
}
\examples{
\dontrun{ cs <- cs_from_bmrb(15060); cs_peek(cs) }
}
