% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{export_star_data}
\alias{export_star_data}
\title{Exports NMR-STAR file to BMRB API server}
\usage{
export_star_data(filename)
}
\arguments{
\item{filename}{filename with correct path}
}
\value{
Temporary tocken to access the file
}
\description{
Exports NMR-STAR file to BMRB API server for data visualization.  This function will return a tocken, which can be used like a pseudo BMRB ID. The tocken will expire after 7 days
}
\examples{
# ent_id <- export_star_data('/nmrdata/hpr.str')
# Exports hpr.str file to BMRB API server and gets a temporary tocken
}
\seealso{
\code{\link{fetch_atom_chemical_shifts}}, \code{\link{fetch_entry_chemical_shifts}} \code{\link{fetch_res_chemical_shifts}}
}

