% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{fetch_atom_chemical_shifts}
\alias{fetch_atom_chemical_shifts}
\title{NMR Chemical shifts list}
\usage{
fetch_atom_chemical_shifts(atom, db = "macromolecules")
}
\arguments{
\item{atom}{==> atom name like CA,CB2}

\item{db}{==> macromolecules, metabolomics}
}
\value{
list of all atom chemical shifts for all BMRB entries as a R data frame
}
\description{
Downloads the full list of chemical shifts from BMRB(www.bmrb.wisc.edu) macromolecular/metabolomics database
}
\examples{
df<-fetch_atom_chemical_shifts('CB2','macromolecules')
df<-fetch_atom_chemical_shifts('C1','metabolomics')
}

