% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{fetch_entry_chemical_shifts}
\alias{fetch_entry_chemical_shifts}
\title{Downloads chemical shift data from BMRB for a given BMRB entry/list of BMRB entries}
\usage{
fetch_entry_chemical_shifts(BMRBidlist)
}
\arguments{
\item{BMRBidlist}{==> sinlge BMRB ID / list of BMRB IDs in csv format exampe:c(17074,17076,17077)}
}
\value{
all available chemical shift data in R data frame
}
\description{
Downloads chemical shift data from BMRB for a given BMRB entry/list of BMRB entries
}
\examples{
df<-fetch_entry_chemical_shifts(15060)
df<-fetch_entry_chemical_shifts(c(17074,17076,17077))
}

