% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{fetchallBMRB}
\alias{fetchallBMRB}
\title{Reads the full chemical shift csv frile from BMRB ftp site. Not recommended unless you have high speed internet connection}
\usage{
fetchallBMRB()
}
\arguments{
\item{No}{parameter is needed}
}
\value{
list of all atom chemical shifts for all BMRB entries as a R data frame
}
\description{
Reads the full chemical shift csv frile from BMRB ftp site. Not recommended unless you have high speed internet connection
}
\examples{
df<-fetchallBMRB()
}

