% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{N15HSQC}
\alias{N15HSQC}
\title{Converts chemical shift data frame into H1-N15 HSQC data frame}
\usage{
N15HSQC(csdf)
}
\arguments{
\item{csdf}{==> chemical shift data frame from fetchBMRB}
}
\value{
1H-N15 chemical shift list on the same row combined using comp index ID and bmrb ID
}
\description{
Converts chemical shift data frame into H1-N15 HSQC data frame
}
\examples{
df<-fetchBMRB('15060')
hsqc<-N15HSQC(df)
}

