% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{TOCSY}
\alias{TOCSY}
\title{Simulates TOCSY spectra for a given entry or a list of entries from BMRB}
\usage{
TOCSY(idlist, interactive = TRUE)
}
\arguments{
\item{idlist}{list of bmrb ids c(17074,17076,17077)}

\item{interactive}{TRUE/FALSE default=TRUE}
}
\value{
plot object
}
\description{
Simulates TOCSY spectra directly from [BMRB](www.bmrb.wisc.edu) database. 'By default it will generate interactive graphics using plotly library
}
\examples{
plot_tocsy<-TOCSY(c(17074,17076,17077))
#Simulates TOCSY spectra for the given 3 entries
plot_tocsy<-TOCSY(c(17074,17076,17077),interactive=FALSE)
# Example to disable interactive plot feature
}
\seealso{
\code{\link{HSQC_15N}} and \code{\link{HSQC_13C}}
}

