% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMRB.R
\name{TOCSY}
\alias{TOCSY}
\title{Simulate TOCSY spectra}
\usage{
TOCSY(idlist, interactive = FALSE)
}
\arguments{
\item{idlist}{==> list of bmrb ids c(17074,17076,17077)}

\item{interactive}{==> TRUE/FALSE default=FALSE}
}
\value{
plot object
}
\description{
Simulates TOCSY spectra directly from BMRB(www.bmrb.wisc.edu) database. Default plot type will be 'scatter'.Peaks from different spectra(entries) can be connected based on residue numbers by specifying plot type as 'line'
}
\examples{
plot_tocsy<-TOCSY(c(17074,17076,17077))
plot_tocsy<-TOCSY(c(17074,17076,17077),interactive=TRUE)
}

